lateshow-misita/requirements.txt
alembic==1.15.2
aniso8601==10.0.0
apturl==0.5.2
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.9.0
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.1.8
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==0.18.2
greenlet==1.1.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
monotonic==1.6
more-itertools==8.10.0
natsort==8.0.2
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==24.2
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pipenv==2024.4.1
platformdirs==4.3.7
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2024.2
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
scour==0.38.2
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
systemd-python==234
ttystatus==0.38
typing_extensions==4.13.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.29.3
virtualenv-clone==0.3.0
wadllib==1.3.6
watchdog==2.1.6
webencodings==0.5.1
Werkzeug==3.1.3
wxPython==4.0.7
xdg==5
xkit==0.0.0
youtube-dl==2021.12.17
zipp==1.0.0
